Huge left turn
I discovered that kickstarter and kicktraq urls are easily translated between each other. This means I can base the script on the AirTable data and look up the kicktraq info one project at a time. This is much more efficient and allows me to do more. Modified the kicktraq processing functions to grab pretty much all the interesting details from that page and modified the airtable function to return everything in the "Data Entry" view.